Move definition package to top
