+++ core/schedule_manager.go
@@ -11,8 +11,8 @@ import (
 	"sync"
 	"time"
 
-	"github.com/jasonjoo2010/goschedule/core/definition"
 	u "github.com/jasonjoo2010/goschedule/core/utils"
+	"github.com/jasonjoo2010/goschedule/definition"
 	"github.com/jasonjoo2010/goschedule/log"
 	"github.com/jasonjoo2010/goschedule/store"
 	"github.com/jasonjoo2010/goschedule/types"
