+++ store/etcdv2/etcd.go
@@ -15,7 +15,7 @@ import (
 	"time"
 
 	etcd "github.com/coreos/etcd/client"
-	"github.com/jasonjoo2010/goschedule/core/definition"
+	"github.com/jasonjoo2010/goschedule/definition"
 	"github.com/jasonjoo2010/goschedule/store"
 	"github.com/sirupsen/logrus"
 )
